agent_video/requirements.txt
openai>=1.59.9
pipecat-ai[daily,deepgram,openai,silero,cartesia]>=0.0.58
python-dotenv>=1.0.1
pydantic>=2.10.6
watchfiles>=1.0.4
restack-ai>=0.0.81 